+++ packages/core/analytics/src/AnalyticsListener.js
@@ -29,6 +29,13 @@ const ContextTypes = {
   onAnalyticsEvent: PropTypes.func,
 };
 
+if (process.env.NODE_ENV !== 'production') {
+  // eslint-disable-next-line no-console
+  console.warn(
+    'The @atlaskit/analytics package has been deprecated. Please use the @atlaskit/analytics-next package instead.',
+  );
+}
+
 class AnalyticsListener extends Component<Props, {}> {
   static defaultProps = {
     match: '*',