+++ packages/core/analytics/src/withAnalytics.js
@@ -18,6 +18,13 @@ type AnalyticsProps = {
   innerRef?: Function,
 };
 
+if (process.env.NODE_ENV !== 'production') {
+  // eslint-disable-next-line no-console
+  console.warn(
+    'The @atlaskit/analytics package has been deprecated. Please use the @atlaskit/analytics-next package instead.',
+  );
+}
+
 /*
 The withAnalytics HOC wraps a component and provides the `fireAnalyticsEvent`
 and `firePrivateAnalyticsEvent` methods to it as props. It contains the logic