+++ packages/media/media-ui/src/__tests__/unit/media-ui-server-side.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
 import * as ReactDOMServer from 'react-dom/server';
 import { getExamplesFor } from '@atlaskit/build-utils/getExamples';
-
-test('media-ui server side rendering', async done => {
+// https://product-fabric.atlassian.net/browse/BUILDTOOLS-282: SSR tests are still timing out in Landkid.
+test.skip('media-ui server side rendering', async done => {
   const examples = await getExamplesFor('media-ui');
   for (const example of examples) {
     const Example = await require(example.filePath).default;