+++ packages/navigation/atlassian-switcher/src/__tests__/unit/map-results-to-switcher-props.ts
@@ -10,6 +10,7 @@ import {
   AvailableSite,
   WorklensProductType,
   AvailableProduct,
+  ProductTopItemVariation,
 } from '../../types';
 
 describe('map-results-to-switcher-props', () => {
@@ -23,6 +24,7 @@ describe('map-results-to-switcher-props', () => {
           disableCustomLinks: false,
           disableRecentContainers: false,
           isDiscoverMoreForEveryoneEnabled: false,
+          productTopItemVariation: ProductTopItemVariation.currentSite,
           xflow: true,
         },
         asCompletedProvider<AvailableProductsResponse>({ sites: [] }),
@@ -48,6 +50,7 @@ describe('map-results-to-switcher-props', () => {
           disableRecentContainers: false,
           isDiscoverMoreForEveryoneEnabled: false,
           xflow: true,
+          productTopItemVariation: ProductTopItemVariation.currentSite,
         },
         asCompletedProvider<AvailableProductsResponse>({ sites: [] }),
       );
@@ -71,6 +74,7 @@ describe('map-results-to-switcher-props', () => {
           disableCustomLinks: false,
           disableRecentContainers: false,
           isDiscoverMoreForEveryoneEnabled: false,
+          productTopItemVariation: ProductTopItemVariation.currentSite,
           xflow: true,
         },
         asCompletedProvider<AvailableProductsResponse>({ sites: [] }),
@@ -93,6 +97,7 @@ describe('map-results-to-switcher-props', () => {
           disableCustomLinks: false,
           disableRecentContainers: false,
           isDiscoverMoreForEveryoneEnabled: false,
+          productTopItemVariation: ProductTopItemVariation.currentSite,
           xflow: false,
         },
         asCompletedProvider<AvailableProductsResponse>({ sites: [] }),
@@ -112,6 +117,7 @@ describe('map-results-to-switcher-props', () => {
           enableUserCentricProducts: true,
           disableCustomLinks: false,
           disableRecentContainers: false,
+          productTopItemVariation: ProductTopItemVariation.currentSite,
           xflow: false,
           isDiscoverMoreForEveryoneEnabled: false,
         },
@@ -181,6 +187,112 @@ describe('map-results-to-switcher-props', () => {
       ]);
     });
 
+    it('shows the most frequently used site at the top of the product if the variation is enabled', () => {
+      const props = mapResultsToSwitcherProps(
+        cloudId,
+        loadingProvidersResult,
+        {
+          enableUserCentricProducts: true,
+          disableCustomLinks: false,
+          disableRecentContainers: false,
+          productTopItemVariation: ProductTopItemVariation.mostFrequentSite,
+          xflow: false,
+          isDiscoverMoreForEveryoneEnabled: false,
+        },
+        asCompletedProvider<AvailableProductsResponse>({
+          sites: [
+            generateSite('site60', [WorklensProductType.JIRA_SOFTWARE, 60]),
+            generateSite('site30', [WorklensProductType.JIRA_SOFTWARE, 30]),
+            generateSite('site10', [WorklensProductType.JIRA_SOFTWARE, 10]),
+            generateSite(cloudId, [WorklensProductType.JIRA_SOFTWARE, 0]),
+          ],
+        }),
+      );
+
+      expect(props.licensedProductLinks).toMatchObject([
+        {
+          description: 'site60',
+          label: 'Jira Software',
+          href: `https://site60.atlassian.net/secure/BrowseProjects.jspa?selectedProjectType=software`,
+          childItems: [
+            {
+              label: 'site10',
+              href:
+                'https://site10.atlassian.net/secure/BrowseProjects.jspa?selectedProjectType=software',
+            },
+
+            {
+              label: 'site30',
+              href:
+                'https://site30.atlassian.net/secure/BrowseProjects.jspa?selectedProjectType=software',
+            },
+            {
+              label: 'site60',
+              href:
+                'https://site60.atlassian.net/secure/BrowseProjects.jspa?selectedProjectType=software',
+            },
+            {
+              label: cloudId,
+              href: `https://${cloudId}.atlassian.net/secure/BrowseProjects.jspa?selectedProjectType=software`,
+            },
+          ],
+        },
+      ]);
+    });
+
+    it('shows the current site at the top of the product by default', () => {
+      const props = mapResultsToSwitcherProps(
+        cloudId,
+        loadingProvidersResult,
+        {
+          enableUserCentricProducts: true,
+          disableCustomLinks: false,
+          disableRecentContainers: false,
+          productTopItemVariation: ProductTopItemVariation.currentSite,
+          xflow: false,
+          isDiscoverMoreForEveryoneEnabled: false,
+        },
+        asCompletedProvider<AvailableProductsResponse>({
+          sites: [
+            generateSite('site60', [WorklensProductType.JIRA_SOFTWARE, 60]),
+            generateSite('site30', [WorklensProductType.JIRA_SOFTWARE, 30]),
+            generateSite('site10', [WorklensProductType.JIRA_SOFTWARE, 10]),
+            generateSite(cloudId, [WorklensProductType.JIRA_SOFTWARE, 0]),
+          ],
+        }),
+      );
+
+      expect(props.licensedProductLinks).toMatchObject([
+        {
+          description: cloudId,
+          label: 'Jira Software',
+          href: `https://${cloudId}.atlassian.net/secure/BrowseProjects.jspa?selectedProjectType=software`,
+          childItems: [
+            {
+              label: 'site10',
+              href:
+                'https://site10.atlassian.net/secure/BrowseProjects.jspa?selectedProjectType=software',
+            },
+
+            {
+              label: 'site30',
+              href:
+                'https://site30.atlassian.net/secure/BrowseProjects.jspa?selectedProjectType=software',
+            },
+            {
+              label: 'site60',
+              href:
+                'https://site60.atlassian.net/secure/BrowseProjects.jspa?selectedProjectType=software',
+            },
+            {
+              label: cloudId,
+              href: `https://${cloudId}.atlassian.net/secure/BrowseProjects.jspa?selectedProjectType=software`,
+            },
+          ],
+        },
+      ]);
+    });
+
     it('shows descriptions for products that belong to multiple sites', () => {
       const props = mapResultsToSwitcherProps(
         cloudId,
@@ -189,6 +301,7 @@ describe('map-results-to-switcher-props', () => {
           enableUserCentricProducts: true,
           disableCustomLinks: false,
           disableRecentContainers: false,
+          productTopItemVariation: ProductTopItemVariation.currentSite,
           xflow: false,
           isDiscoverMoreForEveryoneEnabled: false,
         },
@@ -212,7 +325,7 @@ describe('map-results-to-switcher-props', () => {
       ]);
     });
 
-    it('does not show descriptions for products that belong to the one site', () => {
+    it('does not show descriptions for products that belong to a single site', () => {
       const props = mapResultsToSwitcherProps(
         cloudId,
         loadingProvidersResult,
@@ -221,6 +334,7 @@ describe('map-results-to-switcher-props', () => {
           isDiscoverMoreForEveryoneEnabled: false,
           disableCustomLinks: false,
           disableRecentContainers: false,
+          productTopItemVariation: ProductTopItemVariation.currentSite,
           xflow: false,
         },
         asCompletedProvider<AvailableProductsResponse>({
@@ -250,6 +364,7 @@ describe('map-results-to-switcher-props', () => {
           disableCustomLinks: false,
           disableRecentContainers: false,
           xflow: false,
+          productTopItemVariation: ProductTopItemVariation.currentSite,
           isDiscoverMoreForEveryoneEnabled: false,
         },
         asCompletedProvider<AvailableProductsResponse>({