+++ packages/editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @atlaskit/editor-mobile-bridge
 
+## 8.8.15
+
+- Updated dependencies [ee804f3eeb](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/ee804f3eeb):
+  - @atlaskit/editor-common@39.17.2
+  - @atlaskit/editor-core@112.41.2
+  - @atlaskit/renderer@49.7.7
+  - @atlaskit/media-core@30.0.9
+  - @atlaskit/media-test-helpers@24.3.5
+  - @atlaskit/media-client@2.0.0
+
 ## 8.8.14
 
 - Updated dependencies [688f2957ca](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/688f2957ca):